Files and folder restructuring
